dev-requirements.txt
# git+https://github.com/cloudify-cosmo/cloudify-common@4.4#egg=cloudify-common==4.4

cloudify-dsl-parser==4.3.1
cloudify-rest-client==4.3.1
cloudify-plugins-common==4.3.1

# nose
requests
cerberus

lfm-cli>=2.0.0
# /home/vagrant/lfm-cli